파일 구조를 분리, 명료한 구성으로 변경
Tab 열거형으로 변경

인접해서 사용되는 컴포넌트 그룹화

tab 열거형을 타입으로 사용
