+++ src/components/molecules/InvitationTabs/AfterVisit/index.ts
@@ -1,0 +1,3 @@
+import AfterVisit from "./AfterVisit";
+
+export default AfterVisit;
