+++ src/components/molecules/InvitationTabs/BeforeVisit/index.ts
@@ -1,0 +1,3 @@
+import BeforeVisit from "./BeforeVisit";
+
+export default BeforeVisit;
