+++ src/components/molecules/InvitationTabs/InformationCard/index.ts
@@ -1,0 +1,3 @@
+import InformationCard from "./InformationCard";
+
+export default InformationCard;
