+++ src/components/molecules/InvitationTabs/QRCodeCard/index.ts
@@ -1,0 +1,3 @@
+import QRCodeCard from "./QRCodeCard";
+
+export default QRCodeCard;
