+++ src/components/organizations/GuestbookTabs/index.ts
@@ -1,0 +1,4 @@
+import GuestbookTabs from "./GuestbookTabs";
+
+export type { ITabProps } from "./GuestbookTabs";
+export default GuestbookTabs;
