+++ src/components/organizations/InvitationTabs/index.ts
@@ -1,4 +1,5 @@
 import InvitationTabs from "./InvitationTabs";
 
 export type { ITabProps } from "./InvitationTabs";
+export { Tabs } from "./InvitationTabs";
 export default InvitationTabs;
